backend/pyproject.toml
[project]
name = "npte_react_agent_demo"
version = "0.1.0"
description = "NPTE‑style ReAct MCQA demo (gpt‑4o‑mini, Qdrant, LangSmith)"
requires-python = ">=3.11"

dependencies = [
  "fastapi[standard]>=0.116.1",
  "uvicorn>=0.35.0",
  "langchain-core>=0.3.27",
  "langgraph>=0.2.61",
  "langsmith>=0.4.10",
  "langchain-anthropic>=0.3.18",
  "langchain-openai>=0.2.14",
  "openai>=0.27.0",
  "qdrant-client>=1.7.0",
  "python-dotenv>=1.0",
  "PyPDF2>=3.0.0",
  "python-docx>=0.8.11",
  "langchain>=0.3.14",
  "langchain-cohere==0.4.4",
  "cohere>=5.12.0,<5.13.0", 
  "langsmith>=0.4.10",
  "langchain-ollama>=0.3.6",
  "langchain-qdrant>=0.2.0",
  "ragas>=0.2.10",
  "langchain-community>=0.3.14",
  "langchain-text-splitters>=0.3.9",
  "ollama>=0.5.1",
  "requests>=2.31.0",
  "nltk>=3.8.1",
  "pandas>=2.3.1",
  "datasets>=4.0.0",
  "pymupdf>=1.24.0",
  "unstructured>=0.16.12",
  "jupyter>=1.1.1",
  "rapidfuzz>=3.0.0",
  "pillow>=10.0.0",
]

[tool.uv]
# necessary to treat this as a "package‑project" so uv installs your main as a dependency
package = false
